Introducing the DataExporter SPI.
Data exporters are the classes to get data out of the database for
storage elsewhere. The current plan is to deliver two core data
exporters: One for BOLT requests to server plugins where the request
endpoint and the parameters are given. Another for classes that operate
directly on the graph database by specifying the class name, the method
name and the parameters. Both will need some response consumer that
writes the data to its destination. A third data exporter, namely a
cypher-query exporter, would be a good thing but is not planned for now.