Made better testing.
First, we now use org.testcontainers to get a Neo4j docker container to test in. Second, we use TestNG. The TestSuite class starts the Neo4j container before the tests are run end stops it after.
The current test design is not yet, perfect, however: I figure the VersioningIT should be run in isolation from the other integration tests since it messes around with the version node. Not an issue now.