All tests in the core are running.
Note that all tests are TestNG tests. All integration tests end with
"IT" and use the Neo4j Testcontainers database via Docker.
All IT tests use the ITTestsSetup class that creates the Neo4j DB.
The InserterIT adds a few concepts that the operators and exporters
employ for their tests. The integration-tests.xml file can be used
to run only the integration-tests TestNG suite.