Update JULIE Lab Neo4j plugins to v3.1.0. Adapt tests.
Two tests needed to be adapted because in newer versions, the concept do not have the 'sourceIds' property  but 'sourceIds<number>' for each source ID a concept has. The idea is that the same concept may imported from multiple sources that each may assign another ID. Until version 5, Neo4j did not support array-valued  properties for indexing (now, full-text indices apparently do). So we created separate sourceIds with indices of their own. These are used in the 'lookupConcept' method in the Neo4j plugins ConceptManager. But actually, that method has not been used in production because it was clear most of the time on the application side which ID to use with which node label. Nowadays, I would replace the sourceIDs with labels anyway, I think. But sourceIds are actually older than the labels in Neo4j.